tests: ブラウザテストでReactRendererを表示可能に
